requirements.txt
numpy
pytest
gymnasium
torch
imageio[ffmpeg]
pyyaml
matplotlib
